Ft_printf done

+++ Projects/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: rraumain <rraumain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 08:40:13 by rraumain          #+#    #+#             */
-/*   Updated: 2024/11/05 07:13:10 by rraumain         ###   ########.fr       */
+/*   Updated: 2024/11/05 09:50:57 by rraumain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 
 int	ft_printf(const char *format, ...);
 int	ft_selector(char selector, va_list *args);
-int	ft_puthex(unsigned long n, int is_upper);
+int	ft_puthex(unsigned int n, int is_upper);
 int	ft_putchar(char c);
 int	ft_putnbr(int n);
 int	ft_putstr(char *s);
